lib now lives outside the app project.
I think this makes the project a little clearer. to work you have to have the lib folder setup with the same name as the npm package  in the  the tsconfig. Then you have to have it listed as an dependency with a local install so it becomes a sym link. Then when you run the N application with `tns run ios` (or android) add the flag `--syncAllFiles` so it watchs the sym linked lib folder